rebase: replace mixin approach with constructor, implement setters
It's easier to modify the input to our format with this approach with us being able to add javadoc to the setters.
This also allows us to get rid of the normalize function, since any transformations should happen in the constructor.
Added setters and missing fields, so that all current devices are properly deserialized (Weather and shellies have a different format we have to deal with).