feature: Add yaml for class mapping
Add easy way to add/change topic to class bindings, using regex for more flexibility and allow multiple to map to the same class.
Keep old method as fallback. If this is stable we can remove the legacy/fallback code as well as Battery and PV if they have the same units as Main.