refactor: Remove legacy code to clean up Structure
Remove old method of figuring out the class to a topic name for a more generic and extendable approach.
This allows us to move the getClassFromIdentification() function to a separate class for reusibility.
We can remove unused classes now and also rename them, just make sure to keep the ClassMap.yml up-to-date.