refactor: move javadoc, replace methods, remove dbReader
dbReader isn't worth the effort of sending REST request if we already have the repository and only have basic requests. Since our usage of the writer should stay simple, just parsing data and saving it, we shouldn't rely on another service to be up for that. The reader is where the more demanding requests should be happening anyways.