refactor: replace method implementations from merge
Use hibernate query methods to reproduce the functions of the new methods.
For the more complex aggregations rewrite the sql commands to be used on their own.
Use Null checks and named params to make some parameters optional.
Rewrite queryFlow() more heavily to match the intent of "real-time" data flow.