feat: clean up with the help of lombok annotations
Replace Gettters/Setters with the @Data Annotation.
Use a type def to be able to assign a basic int[] array to psqls int array for compatibility and simplicity of the pre-orm-refactor column.

Replace child classes of Readings with @JsonAlias since it's more understandable/less overhead and hibernate wouldn't save a subclass even if it was casted as an entity.
Use Java Naming Convention even if it's ugly. We don't get much benefit out of using '_' separated names since Jackson can't properly deserialize to that so we will always need a @JsonProperty annotation or a @Column annotation for hibernate.